chore: move helper functions to `utils::halo2`
