+++ halo2-base/src/virtual_region/lookups.rs
@@ -8,7 +8,7 @@ use crate::halo2_proofs::{
     circuit::{Region, Value},
     plonk::{Advice, Column},
 };
-use crate::utils::halo2::raw_assign_advice;
+use crate::utils::halo2::{constrain_virtual_equals_external, raw_assign_advice};
 use crate::{AssignedValue, ContextTag};
 
 use super::copy_constraints::SharedCopyConstraintManager;
@@ -125,8 +125,7 @@ impl<F: Field + Ord, const ADVICE_COLS: usize> VirtualRegionManager<F>
     type Config = Vec<[Column<Advice>; ADVICE_COLS]>;
 
     fn assign_raw(&self, config: &Self::Config, region: &mut Region<F>) {
-        let mut copy_manager =
-            (!self.witness_gen_only).then(|| self.copy_manager().lock().unwrap());
+        let copy_manager = (!self.witness_gen_only).then(|| self.copy_manager().lock().unwrap());
         let cells_to_lookup = self.cells_to_lookup.lock().unwrap();
         // Copy the cells to the config columns, going left to right, then top to bottom.
         // Will panic if out of rows
@@ -140,8 +139,8 @@ impl<F: Field + Ord, const ADVICE_COLS: usize> VirtualRegionManager<F>
             for (advice, &column) in advices.iter().zip(config[lookup_col].iter()) {
                 let bcell =
                     raw_assign_advice(region, column, lookup_offset, Value::known(advice.value));
-                if let Some(copy_manager) = copy_manager.as_mut() {
-                    copy_manager.constrain_virtual_equals_external(region, *advice, bcell.cell());
+                if let Some(copy_manager) = copy_manager.as_ref() {
+                    constrain_virtual_equals_external(region, *advice, bcell.cell(), copy_manager);
                 }
             }
